CVSL-2032 fixing test
The tests would only pass if MergeOffender was the first integration test to run. This was due to declaring a mock bean in MergeOffender test. This dirtied the application context which triggered the application restarting before the test ran.

I think events get sent to the previously registered event listeners not the event listeners linked to the new application context which means the Done never gets executed - based on [this](https://stackoverflow.com/questions/45587213/using-mockbean-in-tests-forces-reloading-of-application-context) stackoverflow.

By registering the mock bean at the application context level this means the application context doesn't need to be rebuilt which fixes the issue and speeds up the tests